# improve baseline survival performance