```
from lifelines import CoxPHFitter
from lifelines.datasets import load_rossi

rossi = load_rossi()
cp = CoxPHFitter()


%timeit cp.fit(rossi, 'week', 'arrest', strata=['race', 'paro', 'mar', 'wexp'])
# with statra: 529 ms vs 636 ms

%timeit cp.fit(rossi, 'week', 'arrest')
# no strata: 38 ms vs 54.5 ms

```


part of #271